+++ src/dbapi/driver/ftds64/freetds/odbc/connectparams.c
@@ -338,7 +338,7 @@ tdoParseProfile(const char *option, const char *value, void *param)
 	if (strcasecmp(p->entry, option) == 0) {
 		tds_strlcpy(p->buffer, value, p->buffer_len);
 
-		p->ret_val = strlen(p->buffer);
+        p->ret_val = (int) strlen(p->buffer);
 		p->found = 1;
 	}
 }
@@ -393,7 +393,7 @@ SQLGetPrivateProfileString(LPCSTR pszSection, LPCSTR pszEntry, LPCSTR pszDefault
 	if (pszDefault && !param.found) {
 		tds_strlcpy(pRetBuffer, pszDefault, nRetBuffer);
 
-		param.ret_val = strlen(pRetBuffer);
+        param.ret_val = (int) strlen(pRetBuffer);
 	}
 
 	fclose(hFile);
